Refactor file directories and rename variables or files
