+++ src/components/ExploreButton/ExploreButton.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import { FiArrowRightCircle } from 'react-icons/fi';
-import { GoButton, ButtonText } from './ExploreButton.styles';
+import { Explore, ButtonText } from './ExploreButton.styles';
 
 const ExploreButton = ({ handleSubmit } = {}) => (
   <div>
-    <GoButton
+    <Explore
       to={handleSubmit}
     >
       <ButtonText>Explore</ButtonText>
       <FiArrowRightCircle />
-    </GoButton>
+    </Explore>
   </div>
 );
