+++ src/components/ExploreButton/ExploreButton.styles.jsx
@@ -1,8 +1,8 @@
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
-import theme from '../../../config/theme';
+import theme from '../../config/theme';
 
-const GoButton = styled(Link)`
+export const Explore = styled(Link)`
   font-size: 50px;
   font-weight: bold;
   display: inline-flex;
@@ -13,11 +13,9 @@ const GoButton = styled(Link)`
     opacity: 0.8;
   }
 `;
-const ButtonText = styled.span`
+export const ButtonText = styled.span`
   font-size: 1rem;
   margin-right: 10px;
 `;
 
-export {
-  GoButton, ButtonText
-};
+
