+++ src/components/PriceSlider/PriceSlider.jsx
