+++ src/components/PriceSlider/PriceSlider.styles.jsx
