+++ src/config/constants.js
@@ -1,6 +1,3 @@
-// Constants
-
-/* LandingFilter options */
 export const regionOptions = [
   { key: 'North', text: 'North', value: 'North' },
   { key: 'South', text: 'South', value: 'South' },
@@ -9,10 +6,13 @@ export const regionOptions = [
   { key: 'Central', text: 'Central', value: 'Central' }
 ];
 
+// Value is set as integer so that the backend will receive just a number
+// and return search results that can accomodate at least the higher range
+// of pax
 export const paxOptions = [
-  { key: '1-5', text: '1-5', value: '1-5' },
-  { key: '6-10', text: '6-10', value: '6-10' },
-  { key: '> 11', text: '> 11', value: '> 11' }
+  { key: '1-5', text: '1-5', value: 5 },
+  { key: '6-10', text: '6-10', value: 10 },
+  { key: '> 11', text: '> 11', value: 11 }
 ];
 
 export const ratingOptions = [
