+++ src/pages/AboutPage/AboutPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import PortalLayout from '../layouts/PortalLayout';
+import PortalLayout from '../../layouts/PortalLayout';
 
 const App = ({ location: { pathname } = {} } = {}) => (
   <PortalLayout pathname={pathname}>TODO: About Page</PortalLayout>
