+++ src/pages/BookmarksPage/BookmarksPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import PortalLayout from '../layouts/PortalLayout';
+import PortalLayout from '../../layouts/PortalLayout';
 
 const Bookmarks = ({ location: { pathname } = {} } = {}) => (
   <PortalLayout pathname={pathname}>TODO: Bookmark</PortalLayout>
