+++ src/pages/DetailsPage/Details.styles.jsx
