--- src/pages/NotFound.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import main from '../images/main.svg';
-
-const Notfound = () => (
-  <img className="ui  centered massive image" src={main} alt="main" />
-);
-
-export default Notfound;
