+++ src/pages/NotFoundPage/NotFoundPage.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import main from '../../images/main.svg';
+
+const Notfound = () => (
+  <img className="ui  centered massive image" src={main} alt="Page Not Found" />
+);
+
+export default Notfound;
